+++ pyfda/plot_widgets/plot_impz.py
@@ -251,7 +251,11 @@ class Plot_Impz(QWidget):
                 self.impz()
 
             elif 'view_changed' in dict_sig:
-                self.draw()
+                if dict_sig['view_changed'] == 'f_S':
+                    self.ui.recalc_freqs()
+                    self.draw()
+                else:
+                    self.draw()
 
             elif 'ui_changed' in dict_sig:
                 # exclude those ui elements  / events that don't require a recalculation
@@ -443,7 +447,6 @@ class Plot_Impz(QWidget):
             self.x = self.ui.A1 * np.sin(2*pi * self.n * self.ui.f1 + phi1) +\
                 self.ui.A2 * np.sin(2*pi * self.n * self.ui.f2 + phi2)
             self.title_str += r'Sinusoidal Signal '
-            self.stim_wdg = ["Amp1", "Amp2", "Phi1","Phi2", "Freq1", "Freq2"]
 
         elif self.ui.stim == "Sinc":
             self.x = self.ui.A1 * sinc(2 * (self.n - self.ui.N//2) * self.ui.f1 + phi1) +\
@@ -473,10 +476,10 @@ class Plot_Impz(QWidget):
 
         elif self.ui.stim == "Rect":
             if self.ui.chk_stim_bl.isChecked():
-                self.x = self.ui.A1 * rect_bl(2*pi * self.n * self.ui.f1 + phi1, duty=0.5)
+                self.x = self.ui.A1 * rect_bl(2*pi * self.n * self.ui.f1 + phi1, duty=self.ui.stim_par1)
                 self.title_str += r'Bandlimited Rect. Signal'
             else:
-                self.x = self.ui.A1 * sig.square(2*pi * self.n * self.ui.f1 + phi1, duty=0.5)
+                self.x = self.ui.A1 * sig.square(2*pi * self.n * self.ui.f1 + phi1, duty=self.ui.stim_par1)
                 self.title_str += r'Rect. Signal'
 
         elif self.ui.stim == "Comb":
@@ -681,7 +684,7 @@ class Plot_Impz(QWidget):
         self.ui.lblFreqUnit1.setText(to_html(f_unit, frmt=unit_frmt))
         self.ui.lblFreqUnit2.setText(to_html(f_unit, frmt=unit_frmt))
         self.t = self.n * fb.fil[0]['T_S']
-        self.ui.load_fs()
+#        self.ui.load_fs()
 
         self.scale_i = self.scale_o = 1
         self.fx_min = -1.