+++ pyfda/plot_widgets/plot_impz_ui.py
@@ -12,6 +12,7 @@ Create the UI for the PlotImz class
 import logging
 logger = logging.getLogger(__name__)
 
+import collections
 from pyfda.libs.compat import (QCheckBox, QWidget, QComboBox, QLineEdit, QLabel, QPushButton,
                       QHBoxLayout, QVBoxLayout, QGridLayout, pyqtSignal, QEvent, Qt)
 
@@ -79,6 +80,7 @@ class PlotImpz_UI(QWidget):
         self.noise = 'none'
         self.DC = 0.0
         self.stim_formula = "A1 * abs(sin(2 * pi * f1 * n))"
+        self.stim_par1 = 0.5
 
         # frequency
         self.plt_freq_resp = "Line"
@@ -95,6 +97,25 @@ class PlotImpz_UI(QWidget):
         self.window_name = "Rectangular"
 
         self.f_scale = fb.fil[0]['f_S']
+
+        self.stim_wdg_dict = collections.OrderedDict()
+        self.stim_wdg_dict.update(
+        {"None":    {"dc", "noise"},
+         "Impulse": {"dc", "a1", "scale", "noise"},
+         "Step":    {"a1", "noise"},
+         "StepErr": {"a1", "noise"},
+         "Cos":     {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"},
+         "Sine":    {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"},
+         "Sinc":    {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"},
+         "Chirp":   {"dc", "a1", "phi1", "f1", "f2", "noise"},
+         "Triang":  {"dc", "a1", "phi1", "f1", "noise", "bl"},
+         "Saw":     {"dc", "a1", "phi1", "f1", "noise", "bl"},
+         "Rect":    {"dc", "a1", "phi1", "f1", "noise", "bl", "par1"},
+         "Comb":    {"dc", "a1", "phi1", "f1", "noise"},
+         "AM":      {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"},
+         "PM / FM": {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"},
+         "Formula": {"dc", "a1", "a2", "phi1", "phi2", "f1", "f2", "noise"}
+         })
 
         self._construct_UI()
         self._enable_stim_widgets()
@@ -202,7 +223,8 @@ class PlotImpz_UI(QWidget):
         self.cmb_plt_time_stmq = QComboBox(self)
         self.cmb_plt_time_stmq.addItems(plot_styles_list)
         qset_cmb_box(self.cmb_plt_time_stmq, self.plt_time_stmq)
-        self.cmb_plt_time_stmq.setToolTip("<span>Plot style for <em>fixpoint</em> (quantized) stimulus.</span>")
+        self.cmb_plt_time_stmq.setToolTip("<span>Plot style for <em>fixpoint</em> "
+                                          "(quantized) stimulus.</span>")
 
         lbl_plt_time_resp = QLabel(to_html("&nbsp;&nbsp;Response y", frmt='bi'), self)
         self.cmb_plt_time_resp = QComboBox(self)
@@ -213,7 +235,8 @@ class PlotImpz_UI(QWidget):
         lbl_win_time = QLabel(to_html("&nbsp;&nbsp;FFT Window", frmt='bi'), self)
         self.chk_win_time = QCheckBox(self)
         self.chk_win_time.setObjectName("chk_win_time")
-        self.chk_win_time.setToolTip('<span>Show FFT windowing function (can be modified in the "Frequency" tab).</span>')
+        self.chk_win_time.setToolTip('<span>Show FFT windowing function (can be '
+                                     'modified in the "Frequency" tab).</span>')
         self.chk_win_time.setChecked(False)
 
         lbl_log_time = QLabel(to_html("dB", frmt='b'), self)
@@ -249,14 +272,16 @@ class PlotImpz_UI(QWidget):
         self.lbl_nfft_spgr_time.setVisible(spgr_en)
         self.led_nfft_spgr_time = QLineEdit(self)
         self.led_nfft_spgr_time.setText(str(self.nfft_spgr_time))
-        self.led_nfft_spgr_time.setToolTip("<span>Number of FFT points per spectrogram segment.</span>")
+        self.led_nfft_spgr_time.setToolTip("<span>Number of FFT points per "
+                                           "spectrogram segment.</span>")
         self.led_nfft_spgr_time.setVisible(spgr_en)
 
         self.lbl_ovlp_spgr_time = QLabel(to_html("&nbsp;N_OVLP =", frmt='bi'), self)
         self.lbl_ovlp_spgr_time.setVisible(spgr_en)
         self.led_ovlp_spgr_time = QLineEdit(self)
         self.led_ovlp_spgr_time.setText(str(self.ovlp_spgr_time))
-        self.led_ovlp_spgr_time.setToolTip("<span>Number of overlap data points between spectrogram segments.</span>")
+        self.led_ovlp_spgr_time.setToolTip("<span>Number of overlap data points "
+                                           "between spectrogram segments.</span>")
         self.led_ovlp_spgr_time.setVisible(spgr_en)
 
         self.lbl_mode_spgr_time = QLabel(to_html("&nbsp;Mode", frmt='bi'), self)
@@ -274,7 +299,8 @@ class PlotImpz_UI(QWidget):
         self.lbl_byfs_spgr_time.setVisible(spgr_en)
         self.chk_byfs_spgr_time = QCheckBox(self)
         self.chk_byfs_spgr_time.setObjectName("chk_log_spgr")
-        self.chk_byfs_spgr_time.setToolTip("<span>Display spectral density i.e. scale by f_S</span>")
+        self.chk_byfs_spgr_time.setToolTip("<span>Display spectral density i.e. "
+                                           "scale by f_S</span>")
         self.chk_byfs_spgr_time.setChecked(True)
         self.chk_byfs_spgr_time.setVisible(spgr_en)
 
@@ -352,7 +378,8 @@ class PlotImpz_UI(QWidget):
         self.cmb_plt_freq_stmq = QComboBox(self)
         self.cmb_plt_freq_stmq.addItems(plot_styles_list)
         qset_cmb_box(self.cmb_plt_freq_stmq, self.plt_freq_stmq)
-        self.cmb_plt_freq_stmq.setToolTip("<span>Plot style for <em>fixpoint</em> (quantized) stimulus.</span>")
+        self.cmb_plt_freq_stmq.setToolTip("<span>Plot style for <em>fixpoint</em> "
+                                          "(quantized) stimulus.</span>")
 
         lbl_plt_freq_resp = QLabel(to_html("&nbsp;Response Y", frmt='bi'), self)
         self.cmb_plt_freq_resp = QComboBox(self)
@@ -465,10 +492,15 @@ class PlotImpz_UI(QWidget):
 
         self.lblStimulus = QLabel(to_html("Type", frmt='bi'), self)
         self.cmbStimulus = QComboBox(self)
-        self.cmbStimulus.addItems(["None","Impulse","Step","StepErr","Cos","Sine", "Sinc", "Chirp",
-                                   "Triang","Saw","Rect","Comb","AM","PM / FM","Formula"])
+        self.cmbStimulus.addItems(self.stim_wdg_dict.keys())
         self.cmbStimulus.setToolTip("Stimulus type.")
         qset_cmb_box(self.cmbStimulus, self.stim)
+
+        self.lblStimPar1 = QLabel(to_html("&alpha; =", frmt='b'), self)
+        self.ledStimPar1 = QLineEdit(self)
+        self.ledStimPar1.setText("0.5")
+        self.ledStimPar1.setToolTip("Duty Cycle, 0 ... 1")
+        self.ledStimPar1.setObjectName("ledStimPar1")
 
         self.chk_stim_bl = QCheckBox("BL", self)
         self.chk_stim_bl.setToolTip("<span>The signal is bandlimited to the Nyquist frequency "
@@ -498,6 +530,8 @@ class PlotImpz_UI(QWidget):
         layHCmbStim = QHBoxLayout()
         layHCmbStim.addWidget(self.cmbStimulus)
         layHCmbStim.addWidget(self.chk_stim_bl)
+        layHCmbStim.addWidget(self.lblStimPar1)
+        layHCmbStim.addWidget(self.ledStimPar1)
         layHCmbStim.addWidget(self.chk_scale_impz_f)
         layHCmbStim.addWidget(self.cmbChirpMethod)
 
@@ -663,6 +697,7 @@ class PlotImpz_UI(QWidget):
         self.cmbChirpMethod.currentIndexChanged.connect(self._update_chirp_method)
         self.ledDC.editingFinished.connect(self._update_DC)
         self.ledStimFormula.editingFinished.connect(self._update_stim_formula)
+        self.ledStimPar1.editingFinished.connect(self._update_stim_par1)
 
 #------------------------------------------------------------------------------
     def eventFilter(self, source, event):
@@ -678,6 +713,8 @@ class PlotImpz_UI(QWidget):
         - When a QLineEdit widget loses input focus (``QEvent.FocusOut``), store
           current value normalized to f_S with full precision (only if
           ``spec_edited == True``) and display the stored value in selected format
+
+          Emit 'ui_changed':'stim'
         """
 
         def _store_entry(source):
@@ -707,7 +744,7 @@ class PlotImpz_UI(QWidget):
         if event.type() in {QEvent.FocusIn,QEvent.KeyPress, QEvent.FocusOut}:
             if event.type() == QEvent.FocusIn:
                 self.spec_edited = False
-                self.load_fs()
+                self.update_freqs()
             elif event.type() == QEvent.KeyPress:
                 self.spec_edited = True # entry has been changed
                 key = event.key()
@@ -735,63 +772,80 @@ class PlotImpz_UI(QWidget):
 
 
     def _enable_stim_widgets(self):
-        """ Enable / disable widgets depending on the selected stimulus"""
+        """ Enable / disable widgets depending on the selected stimulus """
         self.stim = qget_cmb_box(self.cmbStimulus, data=False)
-        f1_en = self.stim in {"Cos","Sine","Sinc","Chirp","PM / FM","AM","Formula","Rect","Saw","Triang","Comb"}
-        f2_en = self.stim in {"Cos","Sine","Sinc","Chirp","PM / FM","AM","Formula"}
-        dc_en = self.stim not in {"Step", "StepErr"}
+        stim_wdg = self.stim_wdg_dict[self.stim]
 
-        self.chk_stim_bl.setVisible(self.stim in {"Triang", "Saw", "Rect"})
+        self.lblDC.setVisible("dc" in stim_wdg)
+        self.ledDC.setVisible("dc" in stim_wdg)
 
-        self.lblAmp1.setVisible(self.stim != "None")
-        self.ledAmp1.setVisible(self.stim != "None")
         self.chk_scale_impz_f.setVisible(self.stim == 'Impulse')
-        self.chk_scale_impz_f.setEnabled((self.noi == 0 or self.cmbNoise.currentText() == 'None')\
-                                         and self.DC == 0)
+        self.chk_scale_impz_f.setEnabled(self.DC == 0 and (self.noi == 0 or\
+            self.cmbNoise.currentText() == 'None'))
 
-        self.cmbChirpMethod.setVisible(self.stim == 'Chirp')
+        self.lblStimPar1.setVisible("par1" in stim_wdg)
+        self.ledStimPar1.setVisible("par1" in stim_wdg)
 
-        self.lblPhi1.setVisible(f1_en)
-        self.ledPhi1.setVisible(f1_en)
-        self.lblPhU1.setVisible(f1_en)
-        self.lblFreq1.setVisible(f1_en)
-        self.ledFreq1.setVisible(f1_en)
-        self.lblFreqUnit1.setVisible(f1_en)
+        self.chk_stim_bl.setVisible("bl" in stim_wdg)
 
-        self.lblFreq2.setVisible(f2_en)
-        self.ledFreq2.setVisible(f2_en)
-        self.lblFreqUnit2.setVisible(f2_en)
-        self.lblAmp2.setVisible(f2_en and self.stim != "Chirp")
-        self.ledAmp2.setVisible(f2_en and self.stim != "Chirp")
-        self.lblPhi2.setVisible(f2_en and self.stim != "Chirp")
-        self.ledPhi2.setVisible(f2_en and self.stim != "Chirp")
-        self.lblPhU2.setVisible(f2_en and self.stim != "Chirp")
+        self.lblAmp1.setVisible("a1" in stim_wdg)
+        self.ledAmp1.setVisible("a1" in stim_wdg)
+        self.lblPhi1.setVisible("phi1" in stim_wdg)
+        self.ledPhi1.setVisible("phi1" in stim_wdg)
+        self.lblPhU1.setVisible("phi1" in stim_wdg)
+        self.lblFreq1.setVisible("f1" in stim_wdg)
+        self.ledFreq1.setVisible("f1" in stim_wdg)
+        self.lblFreqUnit1.setVisible("f1" in stim_wdg)
 
-        self.lblDC.setVisible(dc_en)
-        self.ledDC.setVisible(dc_en)
+        self.lblAmp2.setVisible("a2" in stim_wdg)
+        self.ledAmp2.setVisible("a2" in stim_wdg)
+        self.lblPhi2.setVisible("phi2" in stim_wdg)
+        self.ledPhi2.setVisible("phi2" in stim_wdg)
+        self.lblPhU2.setVisible("phi2" in stim_wdg)
+        self.lblFreq2.setVisible("f2" in stim_wdg)
+        self.ledFreq2.setVisible("f2" in stim_wdg)
+        self.lblFreqUnit2.setVisible("f2" in stim_wdg)
 
         self.lblStimFormula.setVisible(self.stim == "Formula")
         self.ledStimFormula.setVisible(self.stim == "Formula")
 
+        self.cmbChirpMethod.setVisible(self.stim == 'Chirp')
+
         self.sig_tx.emit({'sender':__name__, 'ui_changed':'stim'})
 
 #-------------------------------------------------------------
-    def load_fs(self):
+    def recalc_freqs(self):
         """
-        Reload sampling frequency from filter dictionary and transform
-        the displayed frequency spec input fields according to the units
-        setting (i.e. f_S). Spec entries are always stored normalized w.r.t. f_S
-        in the dictionary; when f_S or the unit are changed, only the displayed values
-        of the frequency entries are updated, not the dictionary!
+        Update normalized frequencies if required. This is called by via signal
+        ['ui_changed':'f_S'] from plot_impz.process_sig_rx
+        """
+        if fb.fil[0]['freq_locked']:
+            self.f1 *= fb.fil[0]['f_S_prev'] / fb.fil[0]['f_S']
+            self.f2 *= fb.fil[0]['f_S_prev'] / fb.fil[0]['f_S']
 
-        The value for f_scale (which is equal to f_S except when the frequency
-        unit is k) is updated in the draw() routine of plot_impz.py
+        self.update_freqs()
 
-        load_fs() is called during init and when the frequency unit or the
-        sampling frequency have been changed.
+        self.sig_tx.emit({'sender':__name__, 'ui_changed':'f1_f2'})
 
-        It should be called when sigSpecsChanged or sigFilterDesigned is emitted
-        at another place, indicating that a reload is required.
+#-------------------------------------------------------------
+    def update_freqs(self):
+        """
+        `update_freqs()` is called:
+
+        - when one of the stimulus frequencies has changed via eventFilter()
+        - sampling frequency has been changed via signal ['ui_changed':'f_S']
+          from plot_impz.process_sig_rx -> self.recalc_freqs
+
+        The sampling frequency is loaded from filter dictionary and stored as
+        `self.f_scale` (except when the frequency unit is k when `f_scale = self.N`).
+
+        Frequency field entries are always stored normalized w.r.t. f_S in the
+        dictionary: When the `f_S` lock button is unlocked, only the displayed
+        values for frequency entries are updated with f_S, not the dictionary.
+
+        When the `f_S` lock button is pressed, the absolute frequency values in
+        the widget fields are kept constant, and the dictionary entries are updated.
+
         """
 
         # recalculate displayed freq spec values for (maybe) changed f_S
@@ -813,6 +867,7 @@ class PlotImpz_UI(QWidget):
             self.ledFreq2.setText(
                 str(params['FMT'].format(self.f2 * self.f_scale)))
 
+#-------------------------------------------------------------
 
     def _update_amp1(self):
         """ Update value for self.A1 from QLineEditWidget"""
@@ -843,6 +898,7 @@ class PlotImpz_UI(QWidget):
         self.chirp_method = qget_cmb_box(self.cmbChirpMethod) # read current data string
         self.sig_tx.emit({'sender':__name__, 'ui_changed':'chirp_method'})
 
+#-------------------------------------------------------------
 
     def _update_noi(self):
         """ Update type + value + label for self.noi for noise"""
@@ -880,6 +936,13 @@ class PlotImpz_UI(QWidget):
         self.stim_formula = self.ledStimFormula.text().strip()
         self.ledStimFormula.setText(str(self.stim_formula))
         self.sig_tx.emit({'sender':__name__, 'ui_changed':'stim_formula'})
+
+    def _update_stim_par1(self):
+        """ Update value for self.par1 from QLineEditWidget"""
+        self.stim_par1 = safe_eval(self.ledStimPar1.text(), self.stim_par1,
+                                   sign = 'pos', return_type='float')
+        self.ledStimPar1.setText(str(self.stim_par1))
+        self.sig_tx.emit({'sender':__name__, 'ui_changed':'stim_par1'})
 
     # -------------------------------------------------------------------------